Tweaks
Changed how the pop-up menus displayed to fix an issue with the score indicators which were being covered up by the tiles. Having the score indicators always be ontop ment that they were appearing ontop of the pop up windows. But when the popup windows were set to a higher layer height the buttons on them stopped working. This was fixed by adding a graphics raycaster to the popup menus alongside the canvas componet. Also setup a alternative level 2 which uses more tiles then level 1. Need to add interactions for the two newer tiles.